# [MRG+1] Optimize MSE criterion by avoiding computing constant terms during split optimization.