In the pull request #5041, @jmschrei proposes to approximate the impurity
improvement during the optimisation by avoiding to compute all the constant
terms. Since this could be done simply by adding only one method, I have decided
to implement this optimisation.

The benchmarks show that:

- GBRT is 35% faster on covertype;
- random forest and decision tree are 10% faster on mnist regression;
- no impact on extra trees / random forest / decision tree classifier;
- trees between master and this pr are different due to differences in numerical errors and in tie breaking.

PS : I haven't been able to benchmark against #5041 since I had segmentation faults.

Here are the raw benchmark results.

```
± make in && python benchmarks/bench_mnist.py --classifiers RandomForest ExtraTrees CART

...

Classifier                          train-time test-time error-rate

master ExtraTrees                   44.73s       0.50s       0.0288
master RandomForest                 42.24s       0.43s       0.0304
master CART                         20.87s       0.02s       0.1214

thispr ExtraTrees                   46.92s       0.50s       0.0294
thispr RandomForest                 41.75s       0.42s       0.0318
thispr CART                         20.56s       0.18s       0.1219
```

```
± python benchmarks/bench_covertype.py --classifiers CART ExtraTrees GBRT RandomForest

...

Classifier                          train-time test-time error-rate

master RandomForest  62.1709s   0.3572s     0.0334
master ExtraTrees    40.9040s   0.5348s     0.0366
master CART          15.2244s   0.0225s     0.0425
master GBRT         682.0862s   0.2816s     0.1777

thispr RandomForest  66.0639s   0.3583s     0.0330
thispr ExtraTrees    43.6342s   0.5493s     0.0372
thispr CART          14.8608s   0.0683s     0.0424
thispr GBRT         507.0393s   0.3322s     0.1777  
```

```
# mnist in regression

from sklearn.datasets import fetch_mldata
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import ExtraTreesRegressor
from sklearn.tree import DecisionTreeRegressor
try:
    from time import process_time as bench_time
except ImportError:
    from time import time as bench_time
data = fetch_mldata('MNIST original')
for name, EstimatorClass in  [("random forest", RandomForestRegressor),
                              ("extra trees", ExtraTreesRegressor),
                              ('cart', DecisionTreeRegressor)]:
    est = EstimatorClass(random_state=0)
    time_start = bench_time()
    est.fit(data['data'], data['target'])
    chrono = bench_time() - time_start
    print('%s = %s' % (name, chrono))


# ± make in && python bench_mnist_regression.py
#
# master random forest = 155.378248
# master extra trees   = 136.87747000000002
# master cart = 25.426823
#
# thispr random forest = 139.21123
# thispr extra trees   = 134.557126
# thispr cart          = 23.304993
```