@arjoly 2015-09-02
The big difference for gbrt could be due to the difference in the computation of impurity improvement.