@jmschrei 2015-09-02
Which version of my code triggered the seg fault? I just pushed the last major code contribution to #5041, which shouldn't produce any seg faults. If so, let me know and I will track it down.