@arjoly 2015-09-02
With the fix, I still have a nice speed up
```
thispr GBRT         507.0393s   0.3322s     0.1777  
````