@jmschrei 2015-09-02
I'm not sure how helpful this is, but here are the results of the benchmark code you provided, comparing master against my PR on my machine (which is apparently very slow!)
```
#5041
random forest = 171.877454996
extra trees = 240.350306988
cart = 28.9209749699

master
random forest = 291.686326981
extra trees = 291.1176579
cart = 49.7908821106
```