@GaelVaroquaux 2015-09-02
> In the pull request #5041, @jmschrei proposes to approximate the
> impurity improvement during the optimisation by avoiding to compute all
> the constant terms. Since this could be done simply by adding only one
> method, I have decided to implement this optimisation.

It seems to me that #5041 proposes additional speed ups (based on deeper
changes to the code).

Is there any reason to merge this PR rather than #5041?