@glouppe 2015-09-03
J'ai refait le benchmark pour MNIST version régression:

```
# master
random forest = 190.507488012
extra trees = 165.347410917
cart = 28.303839922

# arjoly@fast-mse
random forest = 150.614963055
extra trees = 117.137885094
cart = 23.5541858673

# jmschrei@_tree_rewrite
random forest = 159.706566095
extra trees = 215.314203978
cart = 26.6369800568
```