@glouppe 2015-09-03
> I find that benchmark very confusing, since it doesn't match my results at all (particular ExtraTrees being slower). Which version of my code is that with?

4283b209d65e44975d9b4230ab74b4aaddecdd45