@glouppe 2015-09-03
@jmschrei I just run the benchmark once again for your branch:

```
# jmschrei@_tree_rewrite, 2nd run
random forest = 173.878245115
extra trees = 231.97809577
cart = 23.894657135
```

So it seems there is some variance on my machine. One thing for sure is that extra-trees are slower than what they should.