@arjoly 2015-09-03
@jmschrei I ran the benchmark against #5041 (af9c1cb) on my machine

```
# jmschrei@_tree_rewrite #5041 (af9c1cb) - first run
random forest = 161.766252
extra trees = 236.50657800000002
cart = 28.638787000000036

# jmschrei@_tree_rewrite #5041 (af9c1cb) - second run
random forest = 161.42367199999998
extra trees = 235.89488899999998
cart = 28.593887999999993
```