@ogrisel 2015-09-03
Here are the results of the regressors on mnist on my box (Xeon X5660  @ 2.80GHz with 12MB cache):

```python
from sklearn.datasets import fetch_mldata
from sklearn.ensemble import *
from sklearn.tree import *
import time

data = fetch_mldata('MNIST original')
estimators = [
    ("Random Forest", RandomForestRegressor),
    ("Extra Trees", ExtraTreesRegressor),
    ("CART", DecisionTreeRegressor),
    # ("GBRT", GradientBoostingRegressor),
]

for name, estimator in estimators:
    est = estimator(random_state=0)
    tic = time.time()
    est.fit(data.data, data.target)
    toc = time.time() - tic
    print(name, toc)

# #5041
# Random Forest 150.26244401931763
# Extra Trees 216.57464289665222
# CART 26.032639026641846

# #5203
# Random Forest 137.6741919517517
# Extra Trees 112.64128398895264
# CART 22.449392080307007

# master
# Random Forest 264.32571601867676
# Extra Trees 263.0369408130646
# CART 45.76998996734619
```

So indeed @arjoly's refactoring seems to be always the best from a speed point of view. Before merging we need to check that the grown trees are still the same as in master or at least the accuracy is still the same in case random feature sampling has been affected in one way or another. Also we should check that this does not incur a regression from a memory usage point of view with `memory_profiler`.