@ogrisel 2015-09-03
Here are the result for regressors on mnist on my Mac (clang / llvm 6.1.0 compiler, Core i7 @ 2GHz with 4MB cache), 

```python
# master
# Random Forest 182.51679611206055
# Extra Trees 160.08595299720764
# CART 30.50537896156311

# #5203
# Random Forest 177.5755307674408
# Extra Trees 169.99811220169067
# CART 29.187539100646973

# #5041
# Random Forest 193.5168468952179
# Extra Trees 278.5783009529114
# CART 34.431031942367554
```

again #5203 shows an improvement although not as strong as on the Xeon proc with GCC. #5041 is slower than both master and #5203.