@jmschrei 2015-09-03
```

                       #5041             master              #5203
MNIST (50% train/test split)
Random Forest      71.29   0.862      122.53   0.862      62.9   0.863
Extra Trees       101.00   0.890      126.44   0.889      47.9   0.892
Regression Tree    11.79   0.693       22.27   0.693      10.3   0.688

covtype (50% train/test split)
Random Forest      50.38   0.904       72.66   0.904      43.5   0.904
Extra Trees        83.60   0.926       75.14   0.925      37.2   0.925
Regression Tree     7.88   0.819       10.82   0.822      6.22   0.821
```

I am getting the same results as @ogrisel, where #5203 is the fastest. My computer is Xeon E5640 @2.67GHz with 12MB cache and gcc 4.8.4. However, #5203 is not building the same tree as master. Here are the tree sizes for the regression trees:

```
           #5041   master  #5203
MNIST       6857    6857    6869
covtype    40371   40381   40349
```

I support merging this PR over #5041 if this difference can be accounted for.