@arjoly 2015-09-03
@jmschrei If this pr get merged, can I put your name in the what's new for this feature? Without #5041, we would have missed those improvements.

I hope that you will bring in small pull requests all the remaining improvements that you have thought in #5041 (caching total weight_n_samples, total statitics, pre-calculs of (y ** 2, w y **2) for mse, etc).

If you have time, it would be great that you review this pull request.