@ogrisel 2015-09-03
One more data point, same mnist regression benchmark, big modern xeon with 30MB cache:

```
master
('Random Forest', 163.37543892860413)
('Extra Trees', 135.14275002479553)
('CART', 26.503570079803467)

#5203
('Random Forest', 116.93278503417969)
('Extra Trees', 96.99032807350159)
('CART', 19.511062145233154)
```