@glouppe 2015-09-03
FYI, on a random sample of 1000 rows of MNIST,:

```
master GBRT (n_estimators=100, max_depth=3), train time=30.6134212017 
#5041  GBRT (n_estimators=100, max_depth=3), train time=19.3394122124 	
#5203  GBRT (n_estimators=100, max_depth=3), train time=17.5684459209 	
XGBoostClassifier (n_estimators=100, max_depth=3, nthread=1), train time=11.0144069195
``` 

Still somewhat slower than xgboost, but we are closing the gap!

@jmschrei do you have insights about what further improvements would be needed to get up to speed?