@jmschrei 2015-09-03
@arjoly of course, thanks! I do plan on submitting small PRs. It was a mistake on my part to try to redo a core part of sklearn in one go. I hope you'll have time to review them in the coming weeks.

@glouppe I am taking a quick break from decision trees to see if I can contribute anywhere else in sklearn until next week, when I'll start reviewing the code of master more thoroughly. The solution is likely more caching between splits on different levels of the tree, such as wy and wy**2.