@glouppe 2015-09-03
@jmschrei Thanks for your positive attitude! I have to admit that we were a bit uncomfortable with how to deal best your PR. You did an amazing work and came with great ideas, but also wanted to change too many things at the same time. I truly wish this was not done for nothing and that these ideas will eventually land into master.