@arjoly 2015-09-04
I took into account comments an added a entry in the what's new.