@arjoly 2015-09-04
The last optimization of friedman mse allows to speed up a little more gbrt:
```
covertype - this pr
GBRT         498.4850s   0.3465s     0.1777  
```